requirements.txt
langchain>=0.1.0
langchain-core>=0.1.0
google-cloud-aiplatform>=1.36.0
langchain-google-genai>=0.0.5
playwright>=1.41.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.0.0
SpeechRecognition>=3.10.0
PyAudio>=0.2.14; platform_system == "Windows"
pytest>=7.0.0
pytest-playwright>=0.4.0
